backend/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
sqlmodel==0.0.22
python-dateutil==2.9.0.post0
pydantic==2.9.2
pydantic-core==2.23.4
SQLAlchemy==2.0.36
